+++ examples/aci-connector/main.tf
@@ -1,0 +1,100 @@
+provider "azurerm" {
+  features {}
+}
+
+module "rg" {
+  source  = "bcochofel/resource-group/azurerm"
+  version = "1.4.0"
+
+  name     = "rg-${var.identifier}-example"
+  location = var.location
+}
+
+resource "azurerm_user_assigned_identity" "main" {
+  resource_group_name = module.rg.name
+  location            = module.rg.location
+
+  name = "identity-${var.identifier}-example"
+}
+
+module "vnet" {
+  source  = "bcochofel/virtual-network/azurerm"
+  version = "1.2.1"
+
+  resource_group_name = module.rg.name
+  name                = "vnet-${var.identifier}-example"
+  address_space       = ["10.5.0.0/16"]
+
+  depends_on = [module.rg]
+}
+
+module "subnet" {
+  source  = "bcochofel/subnet/azurerm"
+  version = "1.3.1"
+
+  name                 = "snet-${var.identifier}-example"
+  resource_group_name  = module.rg.name
+  virtual_network_name = module.vnet.name
+  address_prefixes     = ["10.5.0.0/21"]
+}
+
+resource "azurerm_subnet" "example-aci" {
+  name                 = "snet-aci-example"
+  resource_group_name  = module.rg.name
+  virtual_network_name = module.vnet.name
+  address_prefixes     = ["10.5.8.0/24"]
+
+  delegation {
+    name = "aciDelegation"
+    service_delegation {
+      name    = "Microsoft.ContainerInstance/containerGroups"
+      actions = ["Microsoft.Network/virtualNetworks/subnets/action"]
+    }
+  }
+}
+
+resource "azurerm_role_assignment" "network" {
+  scope                = module.rg.id
+  role_definition_name = "Network Contributor"
+  principal_id         = azurerm_user_assigned_identity.main.principal_id
+}
+
+module "aks" {
+  source = "../.."
+
+  name                = "aks${var.identifier}"
+  resource_group_name = module.rg.name
+  dns_prefix          = "demolab"
+
+  default_pool_name = "default"
+
+  user_assigned_identity_id = azurerm_user_assigned_identity.main.id
+
+  node_resource_group = "aks-${var.identifier}-example"
+
+  enable_auto_scaling  = true
+  orchestrator_version = "1.18.14"
+  vnet_subnet_id       = module.subnet.id
+  max_count            = 3
+  min_count            = 1
+  node_count           = 1
+
+  enable_aci_connector_linux      = true
+  aci_connector_linux_subnet_name = azurerm_subnet.example-aci.name
+
+  enable_log_analytics_workspace = true
+
+  network_plugin = "azure"
+  network_policy = "calico"
+
+  kubernetes_version = "1.18.14"
+
+  tags = {
+    "ManagedBy" = "Terraform"
+  }
+
+  depends_on = [
+    module.rg,
+    azurerm_role_assignment.network
+  ]
+}